Fix root domain parking -- needs to be "@" instead of empty. Add "www" subdomain. Lower default price.
Currently the instructions tell people to leave the subdomain field blank to park their domain, but this doesn't work anymore. It needs to be `@` to park the root domain. I am just removing the option to specify a subdomain, since it didn't make sense anyway. I also added a `www` A record and lowered the default price to $1000.